requirements.txt
asgiref==3.6.0
Django==4.2
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
gunicorn
psycopg2-binary
whitenoise
django-allauth
djangorestframework
drf-yasg
django-cors-headers
whitenoise
Pillow